refactor: chat[name] 라우트 관련 컴포넌트 디렉토리 수정
