+++ apps/frontend/src/_components/chat/[name]/RecordButton/RecordButton.tsx
