+++ apps/frontend/src/_components/chat/[name]/RecordButton/_error.ts
